+++ nens_auth_client/apps.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 from django.apps import AppConfig
 from django.conf import settings
-
 from nens_auth_client.oauth import oauth
 
 
@@ -20,6 +19,7 @@ class NensAuthClientConfig(AppConfig):
             authorize_url=settings.NENS_AUTH_AUTHORIZE_URL,
             authorize_params=None,
             jwks_uri=settings.NENS_AUTH_JWKS_URI,
+            issuer=settings.NENS_AUTH_ISSUER,
             client_kwargs={"scope": settings.NENS_AUTH_SCOPE},
         )
         return super().ready()